+++ app/Auth/Infrastructure/Models/EloquentUser.php
@@ -2,49 +2,51 @@
 
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App\Auth\Infrastructure\Models;
 
-use Database\Factories\UserFactory;
+use Database\Factories\EloquentUserFactory;
 use Filament\Models\Contracts\FilamentUser;
 use Filament\Panel;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Illuminate\Support\Carbon;
+use Laravel\Sanctum\HasApiTokens;
 
-class User extends Authenticatable implements FilamentUser
+/**
+ * Class EloquentUser
+ *
+ * @property string $id
+ * @property string $name
+ * @property string $email
+ * @property string $password
+ *
+ * @property Carbon $email_verified_at
+ * @property Carbon $created_at
+ * @property Carbon $updated_at
+ */
+class EloquentUser extends Authenticatable implements FilamentUser
 {
-    /** @use HasFactory<UserFactory> */
-    use HasFactory;
     use Notifiable;
     use HasUuids;
+    use HasApiTokens;
+    /** @use  HasFactory<EloquentUserFactory> */
+    use HasFactory;
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var list<string>
-     */
+    protected $table = 'users';
+
     protected $fillable = [
         'name',
         'email',
         'password',
     ];
 
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var list<string>
-     */
     protected $hidden = [
         'password',
         'remember_token',
     ];
 
-    /**
-     * Get the attributes that should be cast.
-     *
-     * @return array<string, string>
-     */
     protected function casts(): array
     {
         return [