+++ app/Filament/Resources/Users/Tables/UsersTable.php
@@ -1,0 +1,72 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\Users\Tables;
+
+use Filament\Actions\ActionGroup;
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Actions\ViewAction;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+
+class UsersTable
+{
+    public static function configure(Table $table): Table
+    {
+        return $table
+            ->columns([
+                TextColumn::make('id')
+                    ->translateLabel()
+                    ->searchable()
+                    ->disabledClick()
+                    ->toggleable()
+                    ->sortable(),
+
+                TextColumn::make('name')
+                    ->translateLabel()
+                    ->searchable()
+                    ->disabledClick()
+                    ->toggleable()
+                    ->sortable(),
+
+                TextColumn::make('email')
+                    ->translateLabel()
+                    ->searchable()
+                    ->disabledClick()
+                    ->toggleable()
+                    ->badge(),
+
+                TextColumn::make('created_at')
+                    ->translateLabel()
+                    ->date('d.m.Y')
+                    ->searchable()
+                    ->toggleable()
+                    ->disabledClick()
+                    ->sortable(),
+
+                TextColumn::make('updated_at')
+                    ->translateLabel()
+                    ->searchable()
+                    ->date('d.m.Y')
+                    ->disabledClick()
+                    ->toggleable(true)
+                    ->sortable(),
+
+
+            ])
+            ->recordActions([
+                ActionGroup::make([
+                    ViewAction::make(),
+                    EditAction::make(),
+                ]),
+            ])
+            ->toolbarActions([
+                BulkActionGroup::make([
+                    DeleteBulkAction::make(),
+                ]),
+            ]);
+    }
+}