+++ config/database.php
@@ -149,7 +149,12 @@ return [
 
         'options' => [
             'cluster' => env('REDIS_CLUSTER', 'redis'),
-            'prefix' => env('REDIS_PREFIX', Str::slug((string) env('APP_NAME', 'laravel')).'-database-'),
+            'prefix' => env(
+                'REDIS_PREFIX',
+                Str::slug(
+                    is_string($app = env('APP_NAME', 'laravel')) ? $app : 'laravel'
+                ) . '-database-'
+            ),
             'persistent' => env('REDIS_PERSISTENT', false),
         ],