+++ config/session.php
@@ -32,7 +32,7 @@ return [
     |
     */
 
-    'lifetime' => (int) env('SESSION_LIFETIME', 120),
+    'lifetime' => is_numeric(env('SESSION_LIFETIME')) ? (int) env('SESSION_LIFETIME') : 120,
 
     'expire_on_close' => env('SESSION_EXPIRE_ON_CLOSE', false),
 
@@ -129,7 +129,7 @@ return [
 
     'cookie' => env(
         'SESSION_COOKIE',
-        Str::slug((string) env('APP_NAME', 'laravel')).'-session'
+        Str::slug(is_string($app = env('APP_NAME', 'laravel')) ? $app : 'laravel')
     ),
 
     /*