+++ routes/web.php
@@ -4,4 +4,6 @@ declare(strict_types=1);
 
 use Illuminate\Support\Facades\Route;
 
-Route::get('/');
+Route::get('/', function () {
+    return view('welcome');
+});